+++ sentry/src/main/java/io/sentry/util/ExceptionUtils.java
@@ -1,7 +1,11 @@
 package io.sentry.util;
 
+import io.sentry.FilterString;
+import java.util.List;
+import java.util.Set;
 import org.jetbrains.annotations.ApiStatus;
 import org.jetbrains.annotations.NotNull;
+import org.jetbrains.annotations.Nullable;
 
 @ApiStatus.Internal
 public final class ExceptionUtils {
@@ -19,5 +23,43 @@ public final class ExceptionUtils {
       rootCause = rootCause.getCause();
     }
     return rootCause;
+  }
+
+  /** Checks if an exception has been ignored. */
+  @ApiStatus.Internal
+  public static @NotNull boolean isIgnored(
+      final @NotNull Set<Class<? extends Throwable>> ignoredExceptionsForType,
+      final @Nullable List<FilterString> ignoredExceptions,
+      final @NotNull Throwable throwable) {
+    if (throwable == null) {
+      return false;
+    }
+
+    final Class<? extends Throwable> throwableClass = throwable.getClass();
+    if (ignoredExceptionsForType.contains(throwableClass)) {
+      return true;
+    }
+
+    if (ignoredExceptions == null || ignoredExceptions.isEmpty()) {
+      return false;
+    }
+    final String throwableClassName = throwableClass.getCanonicalName();
+    if (throwableClassName == null) {
+      return false;
+    }
+
+    for (final FilterString filter : ignoredExceptions) {
+      if (filter.getFilterString().equals(throwableClassName)) {
+        return true;
+      }
+    }
+
+    for (final FilterString filter : ignoredExceptions) {
+      if (filter.matches(throwableClassName)) {
+        return true;
+      }
+    }
+
+    return false;
   }
 }